modify test code

+++ src/std_dev.rs
@@ -1,3 +1,4 @@
+#[allow(dead_code)]
 fn std_dev(ns : Vec<i32>) -> f32 {
     let mn = mean(&ns);
     let mut sum: f32 = 0.0;
@@ -7,6 +8,7 @@ fn std_dev(ns : Vec<i32>) -> f32 {
     (sum / ns.len() as f32).sqrt()
 }
 
+#[allow(dead_code)]
 fn mean(ns: &Vec<i32>) -> f32 {
     let mut sum = 0;
     for n in ns {
@@ -15,6 +17,7 @@ fn mean(ns: &Vec<i32>) -> f32 {
     sum as f32 / ns.len() as f32
 }
 
+#[allow(dead_code)]
 fn pow(f: f32) -> f32 {
     f * f
 }
@@ -24,8 +27,8 @@ mod tests {
     #[test]
     fn test0() {
         let arr = vec![10, 40, 30, 50, 20];
-        let want = 14.1;
-        let got = super::std_dev(arr);
+        let want = 14.1 as u32;
+        let got = super::std_dev(arr) as u32;
         assert_eq!(want, got);
     }
